Move logic from main.cpp -> flit.h, flit.cpp
